fix(036/react): harden k3h4 geometry response typing
- extract K3H4 geometry response types into a dedicated file
- remove fragile inline geometry type block from react api client
- use parser-stable ReadonlyArray forms for geometry payload arrays
- prevent recurring Vite/esbuild parse failures in api.ts hot path
