+++ src/typescript/react/src/lib/api.ts
@@ -3,6 +3,7 @@ type ErrorPayload = {
 };
 
 import type {LaunchGateReasonCode, LaunchGateStatusResponse, LaunchGateVerifyResponse} from './launchGateTypes';
+import type {K3h4EpochGeometryResponse} from './k3h4GeometryTypes.ts';
 import type {NetcodeAbiLayerCoverage, NetcodeAbiLayerKind, NetcodeAbiLayerLedger, NetcodeAbiLayerSnapshot} from '@banana/ui';
 
 type BananaSummaryResponse = {
@@ -1084,42 +1085,6 @@ export async function fetchK3h4ScalingBenchmark(
       '/api/netcode/k3h4/scaling-benchmark',
   );
 }
-
-export type K3h4GeometryPoint2d = {
-  readonly x: number;
-  readonly y: number;
-};
-
-export type K3h4ClusterGeometry2d = {
-  readonly clusterId: number;
-  readonly centroid: K3h4GeometryPoint2d;
-  readonly inscribedRadius: number;
-  readonly polygon: readonly K3h4GeometryPoint2d[];
-};
-
-export type K3h4TokenGeometry2d = {
-  readonly tokenId: string;
-  readonly x: number;
-  readonly y: number;
-  readonly clusterId: number;
-  readonly weightedScore: number;
-};
-
-export type K3h4ProjectionMetadata = {
-  readonly method: string;
-  readonly components: number;
-  readonly explainedVariance: number;
-};
-
-export type K3h4EpochGeometryResponse = {
-  readonly contractVersion: number;
-  readonly sessionId: string;
-  readonly epoch: number;
-  readonly mode: 'multiplicative' | 'power';
-  readonly clusters: readonly K3h4ClusterGeometry2d[];
-  readonly tokens: readonly K3h4TokenGeometry2d[];
-  readonly projectionMetadata: K3h4ProjectionMetadata;
-};
 
 export async function fetchK3h4EpochGeometry(
     baseUrl: string,
